requirements.txt
appdirs
certifi
charset-normalizer
deepdiff
et_xmlfile
idna
lxml
openpyxl
pandas
pillow
PySide6
PySide6_Addons
PySide6_Essentials
python-docx
requests
shiboken6
typing_extensions
urllib3
